+++ tests/testapp/factories.py
@@ -5,6 +5,8 @@ import wagtail_factories
 
 from factory import fuzzy
 from wagtail import blocks
+from wagtail.contrib.redirects.models import Redirect
+from wagtail.models import Site
 
 from testapp.blocks import (
     CustomInterfaceBlock,
@@ -180,3 +182,16 @@ class SimpleModelFactory(factory.django.DjangoModelFactory):
 class MiddlewareModelFactory(factory.django.DjangoModelFactory):
     class Meta:
         model = MiddlewareModel
+
+
+class RedirectFactory(factory.django.DjangoModelFactory):
+    old_path = factory.Faker("slug")
+    # Note: Site needs `hostname` and `port` to be a unique combination,
+    # so this seems like the best option:
+    site = factory.Iterator(Site.objects.all())
+    redirect_page = factory.SubFactory(wagtail_factories.PageFactory)
+    redirect_link = factory.Faker("url")
+    is_permanent = factory.Faker("boolean")
+
+    class Meta:
+        model = Redirect