+++ CHANGELOG.md
@@ -1,8 +1,13 @@
 ## Unreleased
 
+### Added
+
+-   Factory for `Redirect` model in test app ([tba](https://github.com/torchbox/wagtail-grapple/pull/tba)) @jakubmastalerz
+
 ### Fixed
 
 - `test_src_set_invalid_format` not working with Wagtail 5.2 and above. ([#378](https://github.com/torchbox/wagtail-grapple/pull/378)) @JakubMastalerz
+
 
 ## [0.23.0] - 2023-09-29